Don't remove existing package directories automatically.
This was making it too easy to loose work in progress
in "out/respository".  We now ask the user to take care
of this for us.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/22999056

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@885 7dad1e8b-422e-d2af-fbf5-8013b78bd812

+++ build_tools/common.sh
@@ -653,9 +653,16 @@ DefaultExtractStep() {
 
   Banner "Extracting ${ARCHIVE_NAME}"
 
+  if [ -d ${PACKAGE_DIR} ]; then
+    local FULL_DIR=${NACL_PACKAGES_REPOSITORY}/${PACKAGE_DIR}
+    echo "Upstream archive or patch has changed."
+    echo "Please remove existing workspace to continue: '${FULL_DIR}'"
+    exit 1
+  fi
+  Remove ${PACKAGE_DIR}
+
   # make sure the patch step is applied
   rm -f ${NACL_PACKAGES_STAMPDIR}/${PACKAGE_NAME}/patch
-  Remove ${PACKAGE_DIR}
   local EXTENSION="${ARCHIVE_NAME##*.}"
   if [ ${EXTENSION} = "zip" ]; then
     unzip ${ARCHIVE}
